Allow multiple paths on polygon for holes.
Both CoreGraphics and Cairo seem to follow the usual clockwise are
positive and counter-clockwise paths are negative, so I've not tried
to add additional logic here.

At some point it may be desirable to have fill method options to let
the user select between winding and even-odd, but both backends
default to winding count.